refactor(ollama): remove ModelCatalog
