+++ UPGRADE.md
@@ -1,6 +1,20 @@
 UPGRADE FROM 0.6 to 0.7
 =======================
 
+AI Bundle
+---------
+
+ * The `api_catalog` option for `Ollama` has been removed as the catalog is now automatically fetched from the Ollama server
+ * The `api_key` option for `Ollama` is now `null` by default to allow the usage of a `ScopingHttpClient`
+ * The `endpoint` option for `Ollama` is now `null` by default to allow the usage of a `ScopingHttpClient`
+
+Platform
+-----
+
+ * `ModelCatalog` in `Ollama` has been replaced by `OllamaApiCatalog`
+ * `OllamaApiCatalog` in `Ollama` has been replaced to `ModelCatalog`
+ * `Ollama` model is now `final`
+
 Store
 -----
