+++ examples/ollama/stream-with-token-usage.php
@@ -15,7 +15,7 @@ use Symfony\AI\Platform\Message\MessageBag;
 
 require_once dirname(__DIR__).'/bootstrap.php';
 
-$platform = PlatformFactory::create(env('OLLAMA_HOST_URL'), httpClient: http_client());
+$platform = PlatformFactory::create(env('OLLAMA_HOST_URL'), env('OLLAMA_API_KEY'), httpClient: http_client());
 
 $messages = new MessageBag(
     Message::forSystem('You are a helpful assistant.'),
