+++ src/ai-bundle/CHANGELOG.md
@@ -1,6 +1,13 @@
 CHANGELOG
 =========
 
+0.7
+---
+
+ * The `api_catalog` option for `Ollama` has been removed as the catalog is now automatically fetched from the Ollama server
+ * The `api_key` option for `Ollama` is now `null` by default to allow the usage of a `ScopingHttpClient`
+ * The `endpoint` option for `Ollama` is now `null` by default to allow the usage of a `ScopingHttpClient`
+
 0.6
 ---
