+++ src/platform/src/Bridge/Ollama/ModelCatalog.php
@@ -12,212 +12,72 @@
 namespace Symfony\AI\Platform\Bridge\Ollama;
 
 use Symfony\AI\Platform\Capability;
-use Symfony\AI\Platform\Model;
-use Symfony\AI\Platform\ModelCatalog\AbstractModelCatalog;
+use Symfony\AI\Platform\Exception\InvalidArgumentException;
+use Symfony\AI\Platform\ModelCatalog\ModelCatalogInterface;
+use Symfony\Contracts\HttpClient\HttpClientInterface;
 
 /**
  * @author Oskar Stark <oskarstark@googlemail.com>
+ * @author Guillaume Loulier <personal@guillaumeloulier.fr>
  */
-final class ModelCatalog extends AbstractModelCatalog
+final class ModelCatalog implements ModelCatalogInterface
 {
-    /**
-     * @param array<string, array{class: class-string<Model>, capabilities: list<Capability>}> $additionalModels
-     */
-    public function __construct(array $additionalModels = [])
-    {
-        $defaultModels = [
-            'deepseek-r1' => [
-                'class' => Ollama::class,
-                'capabilities' => [
-                    Capability::INPUT_MESSAGES,
-                    Capability::OUTPUT_TEXT,
-                    Capability::OUTPUT_STRUCTURED,
-                    Capability::TOOL_CALLING,
-                ],
-            ],
-            'gpt-oss' => [
-                'class' => Ollama::class,
-                'capabilities' => [
-                    Capability::INPUT_MESSAGES,
-                    Capability::OUTPUT_TEXT,
-                    Capability::OUTPUT_STRUCTURED,
-                    Capability::TOOL_CALLING,
-                ],
-            ],
-            'llama3.1' => [
-                'class' => Ollama::class,
-                'capabilities' => [
-                    Capability::INPUT_MESSAGES,
-                    Capability::OUTPUT_TEXT,
-                    Capability::OUTPUT_STRUCTURED,
-                    Capability::TOOL_CALLING,
-                ],
-            ],
-            'llama3.2' => [
-                'class' => Ollama::class,
-                'capabilities' => [
-                    Capability::INPUT_MESSAGES,
-                    Capability::OUTPUT_TEXT,
-                    Capability::OUTPUT_STRUCTURED,
-                    Capability::TOOL_CALLING,
-                ],
-            ],
-            'llama3' => [
-                'class' => Ollama::class,
-                'capabilities' => [
-                    Capability::INPUT_MESSAGES,
-                    Capability::OUTPUT_TEXT,
-                    Capability::OUTPUT_STRUCTURED,
-                    Capability::TOOL_CALLING,
-                ],
-            ],
-            'mistral' => [
-                'class' => Ollama::class,
-                'capabilities' => [
-                    Capability::INPUT_MESSAGES,
-                    Capability::OUTPUT_TEXT,
-                    Capability::OUTPUT_STRUCTURED,
-                    Capability::TOOL_CALLING,
-                ],
-            ],
-            'qwen3' => [
-                'class' => Ollama::class,
-                'capabilities' => [
-                    Capability::INPUT_MESSAGES,
-                    Capability::OUTPUT_TEXT,
-                    Capability::OUTPUT_STRUCTURED,
-                    Capability::TOOL_CALLING,
-                ],
-            ],
-            'qwen' => [
-                'class' => Ollama::class,
-                'capabilities' => [
-                    Capability::INPUT_MESSAGES,
-                    Capability::OUTPUT_TEXT,
-                    Capability::OUTPUT_STRUCTURED,
-                    Capability::TOOL_CALLING,
-                ],
-            ],
-            'qwen2' => [
-                'class' => Ollama::class,
-                'capabilities' => [
-                    Capability::INPUT_MESSAGES,
-                    Capability::OUTPUT_TEXT,
-                    Capability::OUTPUT_STRUCTURED,
-                    Capability::TOOL_CALLING,
-                ],
-            ],
-            'qwen2.5' => [
-                'class' => Ollama::class,
-                'capabilities' => [
-                    Capability::INPUT_MESSAGES,
-                    Capability::OUTPUT_TEXT,
-                    Capability::OUTPUT_STRUCTURED,
-                    Capability::TOOL_CALLING,
-                ],
-            ],
-            'qwen2.5-coder' => [
-                'class' => Ollama::class,
-                'capabilities' => [
-                    Capability::INPUT_MESSAGES,
-                    Capability::OUTPUT_TEXT,
-                    Capability::OUTPUT_STRUCTURED,
-                    Capability::TOOL_CALLING,
-                ],
-            ],
-            'gemma3n' => [
-                'class' => Ollama::class,
-                'capabilities' => [
-                    Capability::INPUT_MESSAGES,
-                    Capability::OUTPUT_TEXT,
-                    Capability::OUTPUT_STRUCTURED,
-                ],
-            ],
-            'gemma3' => [
-                'class' => Ollama::class,
-                'capabilities' => [
-                    Capability::INPUT_MESSAGES,
-                    Capability::OUTPUT_TEXT,
-                    Capability::OUTPUT_STRUCTURED,
-                ],
-            ],
-            'qwen2.5vl' => [
-                'class' => Ollama::class,
-                'capabilities' => [
-                    Capability::INPUT_MESSAGES,
-                    Capability::OUTPUT_TEXT,
-                    Capability::OUTPUT_STRUCTURED,
-                ],
-            ],
-            'llava' => [
-                'class' => Ollama::class,
-                'capabilities' => [
-                    Capability::INPUT_MESSAGES,
-                    Capability::OUTPUT_TEXT,
-                    Capability::OUTPUT_STRUCTURED,
-                ],
-            ],
-            'phi3' => [
-                'class' => Ollama::class,
-                'capabilities' => [
-                    Capability::INPUT_MESSAGES,
-                    Capability::OUTPUT_TEXT,
-                    Capability::OUTPUT_STRUCTURED,
-                ],
-            ],
-            'gemma2' => [
-                'class' => Ollama::class,
-                'capabilities' => [
-                    Capability::INPUT_MESSAGES,
-                    Capability::OUTPUT_TEXT,
-                    Capability::OUTPUT_STRUCTURED,
-                ],
-            ],
-            'gemma' => [
-                'class' => Ollama::class,
-                'capabilities' => [
-                    Capability::INPUT_MESSAGES,
-                    Capability::OUTPUT_TEXT,
-                    Capability::OUTPUT_STRUCTURED,
-                ],
-            ],
-            'llama2' => [
-                'class' => Ollama::class,
-                'capabilities' => [
-                    Capability::INPUT_MESSAGES,
-                    Capability::OUTPUT_TEXT,
-                    Capability::OUTPUT_STRUCTURED,
-                ],
-            ],
-            'nomic-embed-text' => [
-                'class' => Ollama::class,
-                'capabilities' => [
-                    Capability::INPUT_TEXT,
-                    Capability::INPUT_MULTIPLE,
-                    Capability::EMBEDDINGS,
-                ],
-            ],
-            'bge-m3' => [
-                'class' => Ollama::class,
-                'capabilities' => [
-                    Capability::INPUT_TEXT,
-                    Capability::INPUT_MULTIPLE,
-                    Capability::EMBEDDINGS,
-                ],
-            ],
-            'all-minilm' => [
-                'class' => Ollama::class,
-                'capabilities' => [
-                    Capability::INPUT_TEXT,
-                    Capability::INPUT_MULTIPLE,
-                    Capability::EMBEDDINGS,
-                ],
-            ],
-        ];
+    public function __construct(
+        private readonly HttpClientInterface $httpClient,
+    ) {
+    }
 
-        $this->models = [
-            ...$defaultModels,
-            ...$additionalModels,
-        ];
+    public function getModel(string $modelName): Ollama
+    {
+        $response = $this->httpClient->request('POST', '/api/show', [
+            'json' => [
+                'model' => $modelName,
+            ],
+        ]);
+
+        $payload = $response->toArray();
+
+        if ([] === $payload['capabilities']) {
+            throw new InvalidArgumentException('The model information could not be retrieved from the Ollama API. Your Ollama server might be too old. Try upgrade it.');
+        }
+
+        $capabilities = array_map(
+            static fn (string $capability): Capability => match ($capability) {
+                'embedding' => Capability::EMBEDDINGS,
+                'completion' => Capability::INPUT_MESSAGES,
+                'tools' => Capability::TOOL_CALLING,
+                'thinking' => Capability::THINKING,
+                'vision' => Capability::INPUT_IMAGE,
+                default => throw new InvalidArgumentException(\sprintf('The "%s" capability is not supported', $capability)),
+            },
+            $payload['capabilities'],
+        );
+
+        if (!\in_array(Capability::EMBEDDINGS, $capabilities, true)) {
+            $capabilities[] = Capability::OUTPUT_STRUCTURED;
+        }
+
+        return new Ollama($modelName, $capabilities);
+    }
+
+    public function getModels(): array
+    {
+        $response = $this->httpClient->request('GET', '/api/tags');
+
+        $models = $response->toArray();
+
+        return array_merge(...array_map(
+            function (array $model): array {
+                $retrievedModel = $this->getModel($model['name']);
+
+                return [
+                    $retrievedModel->getName() => [
+                        'class' => Ollama::class,
+                        'capabilities' => $retrievedModel->getCapabilities(),
+                    ],
+                ];
+            },
+            $models['models'],
+        ));
     }
 }
